requirements.txt
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
beautifulsoup4==4.8.1
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
cssselect2==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-crispy-forms==1.7.2
django-currencies==0.9.0
django-heroku==0.3.1
django-model-utils==3.2.0
django-money==0.15.1
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
isort==4.3.21
jsonfield==2.0.2
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.2.0
progressbar2==3.47.0
psycopg2==2.8.3
py-moneyed==0.8.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
PyPDF2==1.26.0
Pyphen==0.9.5
python-utils==2.3.0
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
swapper==1.1.1
tinycss2==1.0.2
tqdm==4.36.1
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.6
WeasyPrint==51
webencodings==0.5.1
whitenoise==3.2
wrapt==1.11.2
